Move the item model to the models folder
